Initial commit
This initial commit contains still unsolved bugs and needs a code rewrite for optimization sake.
This current iteration is usable, but some bugs may happen, specially when using the tuning.XML importer, and usability may feel limited in very specific cases.
Some of the functionality and interfaces may need to be rethinked to improve applicability to a bigger variety of cases.